requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
jupyter>=1.0.0
ipykernel>=6.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
tqdm>=4.65.0

